refactor(front): move reclassified-internal DTO types out of lib/api into types/api
Thirteenth batch of the lib/api DTO migration. These 9 files were initially
deferred as "public" because v1 routes import from them, but symbol-level
analysis showed v1 routes import only their business FUNCTIONS (which stay in
lib/api), never their DTO types. Their DTO types are internal, so they move to
types/api/* like the earlier internal batches.

- Files: apps, assistant/user_relation, dust_app_secrets, files/upload_metadata,
  provider_credentials, tables, user, data_sources, data_source_view.
- All business functions (resolveLegacyDataSourceSpaceId, handleDataSourceSearch,
  upsertTable, getUserWithWorkspaces, checkAppsDeployment, ...) stay in lib/api;
  the v1 routes that import them are untouched.
- 38 consumers updated via per-specifier rewrite; lib/api/run.ts re-export shim
  repointed to types/api/apps.

Partial: data_source_view.ts keeps GetContentNodesOrChildrenRequestBody (it
references SortingParamsCodec, an io-ts codec in lib/api/pagination — a genuine
types->lib gotcha) for a later pass.
