+++ front-api/routes/w/[wId]/dust_app_secrets/index.ts
@@ -1,7 +1,3 @@
-import type {
-  GetDustAppSecretsResponseBody,
-  PostDustAppSecretsResponseBody,
-} from "@app/lib/api/dust_app_secrets";
 import {
   getDustAppSecret,
   getDustAppSecrets,
@@ -9,6 +5,10 @@ import {
 import { DustAppSecretModel } from "@app/lib/models/dust_app_secret";
 import { rateLimiter } from "@app/lib/utils/rate_limiter";
 import logger from "@app/logger/logger";
+import type {
+  GetDustAppSecretsResponseBody,
+  PostDustAppSecretsResponseBody,
+} from "@app/types/api/dust_app_secrets";
 import { encrypt } from "@app/types/shared/utils/encryption";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import {
