+++ front-api/routes/w/[wId]/spaces/[spaceId]/apps/[aId]/runs/[runId]/cancel.ts
@@ -1,7 +1,7 @@
-import type { PostRunCancelResponseBody } from "@app/lib/api/apps";
 import config from "@app/lib/api/config";
 import { AppResource } from "@app/lib/resources/app_resource";
 import logger from "@app/logger/logger";
+import type { PostRunCancelResponseBody } from "@app/types/api/apps";
 import { CoreAPI } from "@app/types/core/core_api";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
