+++ front-api/routes/w/[wId]/spaces/[spaceId]/apps/[aId]/runs/[runId]/index.ts
@@ -1,5 +1,5 @@
-import type { GetRunResponseBody } from "@app/lib/api/apps";
 import { AppResource } from "@app/lib/resources/app_resource";
+import type { GetRunResponseBody } from "@app/types/api/apps";
 import type { AppType } from "@app/types/app";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
