+++ front-api/routes/w/[wId]/spaces/[spaceId]/data_source_views/[dsvId]/index.ts
@@ -1,11 +1,11 @@
 import config from "@app/lib/api/config";
-import type {
-  GetDataSourceViewResponseBody,
-  PatchDataSourceViewResponseBody,
-} from "@app/lib/api/data_source_view";
 import { handlePatchDataSourceView } from "@app/lib/api/data_source_view";
 import { KillSwitchResource } from "@app/lib/resources/kill_switch_resource";
 import logger from "@app/logger/logger";
+import type {
+  GetDataSourceViewResponseBody,
+  PatchDataSourceViewResponseBody,
+} from "@app/types/api/data_source_view";
 import { PatchDataSourceViewSchema } from "@app/types/api/public/spaces";
 import { ConnectorsAPI } from "@app/types/connectors/connectors_api";
 import type { ConnectorType } from "@app/types/data_source";
