+++ front-api/routes/w/[wId]/spaces/[spaceId]/data_source_views/index.ts
@@ -3,15 +3,15 @@ import {
   getDataSourcesUsageByCategory,
   getDataSourceViewsUsageByModelIds,
 } from "@app/lib/api/agent_data_sources";
-import type {
-  GetSpaceDataSourceViewsResponseBody,
-  PostSpaceDataSourceViewsResponseBody,
-} from "@app/lib/api/data_source_view";
 import { augmentDataSourceWithConnectorDetails } from "@app/lib/api/data_sources";
 import { isManaged, isWebsite } from "@app/lib/data_sources";
 import { DataSourceResource } from "@app/lib/resources/data_source_resource";
 import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { KillSwitchResource } from "@app/lib/resources/kill_switch_resource";
+import type {
+  GetSpaceDataSourceViewsResponseBody,
+  PostSpaceDataSourceViewsResponseBody,
+} from "@app/types/api/data_source_view";
 import { ContentSchema } from "@app/types/api/internal/spaces";
 import type { DataSourceViewCategory } from "@app/types/api/public/spaces";
 import type { DataSourceViewsWithDetails } from "@app/types/data_source_view";
