+++ front/components/spaces/AddConnectionMenu.tsx
@@ -3,7 +3,6 @@ import { CreateOrUpdateConnectionBigQueryModal } from "@app/components/data_sour
 import { CreateOrUpdateConnectionSnowflakeModal } from "@app/components/data_source/CreateOrUpdateConnectionSnowflakeModal";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
 import { useSendNotification } from "@app/hooks/useNotification";
-import type { PostDataSourceRequestBody } from "@app/lib/api/data_sources";
 import { useFeatureFlags } from "@app/lib/auth/AuthContext";
 import { useRegionContext } from "@app/lib/auth/RegionContext";
 import {
@@ -24,6 +23,7 @@ import {
   trackEvent,
   withTracking,
 } from "@app/lib/tracking";
+import type { PostDataSourceRequestBody } from "@app/types/api/data_sources";
 import type {
   ConnectorProvider,
   ConnectorType,
