+++ front/lib/api/apps.ts
@@ -5,9 +5,7 @@ import type { SpaceResource } from "@app/lib/resources/space_resource";
 import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import logger from "@app/logger/logger";
 import tracer from "@app/logger/tracer";
-import type { AppType, SpecificationType } from "@app/types/app";
 import { CoreAPI } from "@app/types/core/core_api";
-import type { RunType } from "@app/types/run";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import type { LightWorkspaceType } from "@app/types/user";
@@ -15,44 +13,6 @@ import type { LightWorkspaceType } from "@app/types/user";
 type AppDeploymentCheck = {
   appId: string;
   appHash: string;
-};
-
-export type GetAppsResponseBody = {
-  apps: AppType[];
-};
-
-export type PostAppResponseBody = {
-  app: AppType;
-};
-
-export type GetOrPostAppResponseBody = {
-  app: AppType;
-};
-
-export type GetRunsResponseBody = {
-  runs: RunType[];
-  total: number;
-};
-
-export type PostRunsResponseBody = {
-  run: RunType;
-};
-
-export type GetRunResponseBody = {
-  run: RunType;
-  spec: SpecificationType;
-};
-
-export type GetRunBlockResponseBody = {
-  run: RunType | null;
-};
-
-export type PostRunCancelResponseBody = {
-  success: boolean;
-};
-
-export type GetRunStatusResponseBody = {
-  run: RunType | null;
 };
 
 export async function softDeleteApp(
