+++ front/lib/api/data_sources.ts
@@ -38,11 +38,7 @@ import type { FrontDataSourceDocumentSectionType } from "@app/types/api/public/d
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
 import { DEFAULT_EMBEDDING_PROVIDER_ID } from "@app/types/assistant/models/embedding";
 import type { AdminCommandType } from "@app/types/connectors/admin/cli";
-import type { ConnectorConfiguration } from "@app/types/connectors/configuration";
-import {
-  ConnectorConfigurationTypeSchema,
-  ConnectorsAPI,
-} from "@app/types/connectors/connectors_api";
+import { ConnectorsAPI } from "@app/types/connectors/connectors_api";
 import type { CoreAPIError, CoreAPITable } from "@app/types/core/core_api";
 import { CoreAPI, EMBEDDING_CONFIGS } from "@app/types/core/core_api";
 import type {
@@ -61,12 +57,7 @@ import type {
   DataSourceWithConnectorDetailsType,
   WithConnector,
 } from "@app/types/data_source";
-import {
-  CONNECTOR_PROVIDERS,
-  isDataSourceNameValid,
-} from "@app/types/data_source";
-import type { DataSourceViewType } from "@app/types/data_source_view";
-import type { DocumentType } from "@app/types/document";
+import { isDataSourceNameValid } from "@app/types/data_source";
 import { OAuthAPI } from "@app/types/oauth/oauth_api";
 import type { PlanType } from "@app/types/plan";
 import type { LLMCredentialsType } from "@app/types/provider_credential";
@@ -84,53 +75,7 @@ import type {
 } from "@dust-tt/client";
 import assert from "assert";
 import type { Transaction } from "sequelize";
-import { z } from "zod";
 import { ConversationResource } from "../resources/conversation_resource";
-
-// Contract types for the spaces data source endpoints (Next + Hono).
-
-export const PostDataSourceWithProviderRequestBodySchema = z.object({
-  provider: z.enum(CONNECTOR_PROVIDERS),
-  name: z.string().optional(),
-  configuration: ConnectorConfigurationTypeSchema,
-  connectionId: z.string().optional(), // Required for some providers
-  relatedCredentialId: z.string().optional(), // Required for private integrations
-  extraConfig: z.record(z.string(), z.string()).optional(), // Used by slack private integrations
-});
-
-const PostDataSourceWithoutProviderRequestBodySchema = z.object({
-  name: z.string(),
-  description: z.string().nullable(),
-});
-
-export const PostDataSourceRequestBodySchema = z.union([
-  PostDataSourceWithoutProviderRequestBodySchema,
-  PostDataSourceWithProviderRequestBodySchema,
-]);
-
-export type PostDataSourceRequestBody = z.infer<
-  typeof PostDataSourceRequestBodySchema
->;
-
-export type PostSpaceDataSourceResponseBody = {
-  dataSource: DataSourceType;
-  dataSourceView: DataSourceViewType;
-};
-
-export type GetDataSourceConfigurationResponseBody = {
-  configuration: ConnectorConfiguration;
-};
-
-export type PatchDataSourceConfigurationResponseBody =
-  GetDataSourceConfigurationResponseBody;
-
-export type PostDocumentResponseBody = {
-  document: DocumentType | CoreAPILightDocument;
-};
-
-export type PatchDocumentResponseBody = {
-  document: DocumentType | CoreAPILightDocument;
-};
 
 const CORE_UNKNOWN_DATA_SOURCE_DELETE_ERROR_PREFIX =
   "Failed to delete data source (error: Unknown DataSource: ";
