+++ front/lib/api/files/upload_metadata.ts
@@ -2,14 +2,9 @@ import { shouldStampSandboxRawDelimited } from "@app/lib/api/files/sandbox_raw";
 import { shouldSkipDataSourceIndexing } from "@app/lib/api/files/should_skip_indexing";
 import type {
   AllSupportedFileContentType,
-  FileTypeWithUploadUrl,
   FileUseCase,
   FileUseCaseMetadata,
 } from "@app/types/files";
-
-export interface FileUploadRequestResponseBody {
-  file: FileTypeWithUploadUrl;
-}
 
 export function buildEffectiveUseCaseMetadata({
   contentType,
