+++ front/lib/swr/data_source_view_documents.ts
@@ -1,9 +1,4 @@
 import { useSendNotification } from "@app/hooks/useNotification";
-import type { GetDataSourceViewDocumentResponseBody } from "@app/lib/api/data_source_view";
-import type {
-  PatchDocumentResponseBody,
-  PostDocumentResponseBody,
-} from "@app/lib/api/data_sources";
 import { clientFetch } from "@app/lib/egress/client";
 import { useDataSourceViewContentNodes } from "@app/lib/swr/data_source_views";
 import {
@@ -11,6 +6,11 @@ import {
   useFetcher,
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
+import type { GetDataSourceViewDocumentResponseBody } from "@app/types/api/data_source_view";
+import type {
+  PatchDocumentResponseBody,
+  PostDocumentResponseBody,
+} from "@app/types/api/data_sources";
 import type {
   PatchDataSourceDocumentRequestBody,
   PostDataSourceDocumentRequestBody,
