+++ front/lib/swr/data_source_view_tables.ts
@@ -1,12 +1,4 @@
 import { useSendNotification } from "@app/hooks/useNotification";
-import type {
-  ListTablesResponseBody,
-  SearchTablesResponseBody,
-} from "@app/lib/api/data_source_view";
-import type {
-  GetDataSourceViewTableResponseBody,
-  PatchTableResponseBody,
-} from "@app/lib/api/tables";
 import { clientFetch } from "@app/lib/egress/client";
 import { useDataSourceViewContentNodes } from "@app/lib/swr/data_source_views";
 import {
@@ -15,7 +7,15 @@ import {
   useFetcher,
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
+import type {
+  ListTablesResponseBody,
+  SearchTablesResponseBody,
+} from "@app/types/api/data_source_view";
 import type { PatchDataSourceTableRequestBody } from "@app/types/api/public/data_sources";
+import type {
+  GetDataSourceViewTableResponseBody,
+  PatchTableResponseBody,
+} from "@app/types/api/tables";
 import type { DataSourceViewType } from "@app/types/data_source_view";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
