+++ front/types/api/dust_app_secrets.ts
@@ -1,0 +1,9 @@
+import type { DustAppSecretType } from "@app/types/dust_app_secret";
+
+export type GetDustAppSecretsResponseBody = {
+  secrets: DustAppSecretType[];
+};
+
+export type PostDustAppSecretsResponseBody = {
+  secret: DustAppSecretType;
+};
