+++ front/types/api/files/upload_metadata.ts
@@ -1,0 +1,5 @@
+import type { FileTypeWithUploadUrl } from "@app/types/files";
+
+export interface FileUploadRequestResponseBody {
+  file: FileTypeWithUploadUrl;
+}
