+++ front/types/api/provider_credentials.ts
@@ -1,0 +1,3 @@
+export type GetProvidersCheckResponseBody =
+  | { ok: true }
+  | { ok: false; error: string };
